:package: organiza o projeto no padrao MVC
