+++ README.md
@@ -1,1 +1,3 @@
 # TB-Java-Projeto-Guiado-II
+
+Nesta atividade foi construido um sistema para o gerenciamento dos colaboradores de uma empresa (CRUD), suas áreas e seus gestores, aplicando conceitos de árvore binária e paradigma orientado a objetos.
