+++ src/funcionariosStarlabs/model/ArvoreDeFuncionarios.java
@@ -1,4 +1,5 @@
-package funcionariosStarlabs;
+package funcionariosStarlabs.model;
+
 import java.util.TreeSet;
 
 public class ArvoreDeFuncionarios<T extends FuncionarioBase> {
@@ -15,7 +16,7 @@ public class ArvoreDeFuncionarios<T extends FuncionarioBase> {
     }
 
     public T buscar(String nome) {
-        
+
         for (T funcionarioBase : funcionarios) {
             if (funcionarioBase.getNome().equals(nome)) {
                 return funcionarioBase; // retorno o funcionário encontrado
@@ -38,7 +39,8 @@ public class ArvoreDeFuncionarios<T extends FuncionarioBase> {
     public void mostrarTodos() {
         System.out.println("\n=== Todos os Funcionário ===");
         for (T funcionarioBase : funcionarios) {
-            System.out.println(funcionarioBase.getNome() + " - " + funcionarioBase.getCargo() + " - " + funcionarioBase.getSalario());
+            System.out.println(funcionarioBase.getNome() + " - " + funcionarioBase.getCargo() + " - "
+                    + funcionarioBase.getSalario());
         }
     }
