+++ src/funcionariosStarlabs/model/FuncionarioBase.java
@@ -1,6 +1,6 @@
-package funcionariosStarlabs;
+package funcionariosStarlabs.model;
 
-public abstract class FuncionarioBase implements FuncionarioBaseInterface {
+public abstract class FuncionarioBase implements FuncionarioInterface {
     protected String nome;
     protected String cargo;
     protected double salario;
@@ -17,30 +17,37 @@ public abstract class FuncionarioBase implements FuncionarioBaseInterface {
     public String getNome() {
         return nome;
     }
+
     @Override
     public void setNome(String nome) {
         this.nome = nome;
     }
+
     @Override
     public String getCargo() {
         return cargo;
     }
+
     @Override
     public void setCargo(String cargo) {
         this.cargo = cargo;
     }
+
     @Override
     public double getSalario() {
         return salario;
     }
+
     @Override
     public void setSalario(double salario) {
         this.salario = salario;
     }
+
     @Override
     public String getCpf() {
         return cpf;
     }
+
     @Override
     public void setCpf(String cpf) {
         this.cpf = cpf;
