+++ src/funcionariosStarlabs/model/FuncionarioBaseNode.java
@@ -1,4 +1,4 @@
-package funcionariosStarlabs;
+package funcionariosStarlabs.model;
 
 public class FuncionarioBaseNode {
     FuncionarioBase funcionario;
