+++ src/funcionariosStarlabs/model/FuncionarioInterface.java
@@ -1,12 +1,19 @@
-package funcionariosStarlabs;
+package funcionariosStarlabs.model;
 
-public interface FuncionarioBaseInterface {
+public interface FuncionarioInterface {
     String getNome();
+
     void setNome(String nome);
+
     String getCargo();
+
     void setCargo(String cargo);
+
     double getSalario();
+
     void setSalario(double salario);
+
     String getCpf();
+
     void setCpf(String cpf);
 }
